P3-B B0h-b: HybridRanker crate split (I4) (#149) (#155)
Move the concrete HybridRanker implementation (and its merge helper) from
kotoha-engine-core out to a new kotoha-ranker-hybrid crate. The engine-core
domain layer is now free of Ranker concrete impl, leaving only the trait
plus value types — this matches spec §3.1 and prepares Phase 4 (fcitx5) /
Phase 5 (multi-backend) for swap-in alternative Ranker concrete crates.

The new crate depends on engine-core (for the Ranker trait, ConversionContext,
CandidateUpdate, CancellationToken, and the learning_port driven ports) plus
kotoha-core (for MorphologicalEngine / KanjiBackend / Candidate). LLM feature
flags (mock-backend / llama-cpp / llama-cpp-smoke) and the test-helpers
transitive flag move with the implementation.

bin DI now imports HybridRanker from the new crate. engine-core's lib.rs
no longer re-exports HybridRanker.
