+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/kotoha-storage",
     "crates/kotoha-engine-core",
     "crates/kotoha-engine-adapter",
+    "crates/kotoha-ranker-hybrid",
     "crates/kotoha-engine-ibus",
     "crates/kotoha-bin",
 ]
